+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## v1.0.0-beta.2
 
-### Removed
+### Changed
 
-- Removed the now redundant method option from backend plugin actions ([#443](https://github.com/starfederation/datastar/issues/443)).
+- Renamed the `@clipboard` action to `@copy` ([#464](https://github.com/starfederation/datastar/issues/464)).
 
 ### Fixed
 
 - Fixed dashes not being trimmed from keys when double dashes were used ([#450](https://github.com/starfederation/datastar/issues/450)).
+
+### Removed
+
+- Removed the now redundant method option from backend plugin actions ([#443](https://github.com/starfederation/datastar/issues/443)).